4.21: bloglist expansion, step9

+++ part04/bloglist-api/controllers/blogs.js
@@ -41,8 +41,18 @@ blogsRouter.post('/', async (request, response, next) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  return response.status(204).end();
 });
 
 blogsRouter.put('/:id', async (request, response, next) => {
